+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added support for the processing of additional elements in `RespektContent`. remp/respekt#388
 - Added HTML WYSIWYG editor for URL Parser email generator intro and footer inputs. remp/remp#1416
 - Added extra index to improve performance of dashboard loading. remp/remp#1418
+- Added setup method for `EmbedParser` to preprocessing of thumbnail image. remp/remp#1411
 
 ## Archive